requirements.txt
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
defusedxml==0.7.0rc1
Django==2.2
django-cors-headers==3.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
idna==2.10
install==1.3.3
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
